--- src/pages/api/messages/definitions/interfaces.ts
@@ -1,23 +1,0 @@
-import { PublicPreKey, SerializedFullDirectoryEntry } from "@/utility/serialize/FullDirectoryEntry";
-import { PreKeyType } from '@privacyresearch/libsignal-protocol-typescript';
-import type { NextApiRequest } from 'next'
-
-
-export interface AddMessage {
-    address: string;
-    message: string;
-}
-
-export interface GetMessages {
-    address: string;
-}
-
-export interface RemoveMessage {
-    address: string;
-}
-
-
-// export interface AddPreKeysBody {
-//     address: string;
-//     keys: PublicPreKey[];
-// }
